requirements.txt
pandas
psycopg2-binary
pymgclient
python-dateutil